refactor: 관심사별 파일 분리 및 모듈화
비즈니스 로직과 UI 컴포넌트를 관심사별로 분리하여 코드 구조 개선:

**분리된 관심사별 모듈:**
- state/appState.js - 애플리케이션 상태 관리
- domains/product/productService.js - 상품 관련 비즈니스 로직
- domains/cart/cartService.js - 장바구니 관련 비즈니스 로직
- domains/loyalty/loyaltyService.js - 포인트 시스템 비즈니스 로직
- components/ui.js - UI 컴포넌트 생성 함수들

**개선사항:**
- 코드 가독성 및 유지보수성 향상
- 관심사의 분리로 응집도 증가, 결합도 감소
- 모듈별 책임 명확화
- 87개 테스트 모두 통과 확인
