+++ src/basic/state/appState.js
@@ -1,0 +1,14 @@
+// Application state management
+export function createAppState() {
+	return {
+		bonusPts: 0,
+		stockInfo: null,
+		itemCnt: 0,
+		lastSel: null,
+		sel: null,
+		addBtn: null,
+		totalAmt: 0,
+		cartDisp: null,
+		sum: null
+	};
+}
